+++ R/ao.R
@@ -18,9 +18,9 @@
 #' The default is \code{as.list(1:length(p))}, i.e. each parameter is
 #' optimized separately.
 #' Parameter indices can be members of multiple subsets.
-#' @param optimizer
+#' @param base_optimizer
 #' An \code{optimizer} object, which can be specified via
-#' \code{\link[optimizeR]{set_optimizer}}.
+#' \code{\link[optimizeR]{define_optimizer}}.
 #' The default optimizer is \code{\link[stats]{optim}}.
 #' @param iterations
 #' An \code{integer}, the number of iterations through the parameter indices in
@@ -56,7 +56,7 @@
 #' himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
 #' ao(
 #'   f = himmelblau, p = c(0,0), partition = list(1, 2), iterations = 10,
-#'   optimizer = optimizer_optim(lower = -5, upper = 5, method = "L-BFGS-B")
+#'   base_optimizer = optimizer_optim(lower = -5, upper = 5, method = "L-BFGS-B")
 #' )
 #'
 #' @export
@@ -65,7 +65,7 @@
 
 ao <- function(
     f, p, ..., partition = as.list(1:length(p)),
-    optimizer = optimizer_optim(), iterations = 10, tolerance = 1e-6,
+    base_optimizer = optimizer_optim(), iterations = 10, tolerance = 1e-6,
     print.level = 0, plot = FALSE
 ) {
   if (missing(f) || !is.function(f)) {
@@ -78,10 +78,10 @@ ao <- function(
       !setequal(unlist(partition), seq_along(p))) {
     ao_stop("'partition' must be a list of vectors of indices of 'p'.")
   }
-  if (!inherits(optimizer, "optimizer")) {
+  if (!inherits(base_optimizer, "optimizer")) {
     ao_stop(
-      "'optimizer' must be an object of class 'optimizer'.",
-      "Use 'optimizeR::set_optimizer()' to create such an object."
+      "Input 'base_optimizer' must be an object of class 'optimizer'.",
+      "Use 'optimizeR::define_optimizer()' to create such an object."
     )
   }
   if (length(iterations) != 1 || !is_number(iterations)) {
@@ -143,19 +143,25 @@ ao <- function(
         theta[p_ind] <- theta_small
         theta[-p_ind] <- est[-p_ind]
         out <- f(theta, ...)
-        if (inherits(out, "gradient")) {
-          attr(out, "gradient") <- attr(out, "gradient")[p_ind]
+        if ("gradient" %in% names(attributes(out))) {
+          gradient <- attr(out, "gradient")
+          if (is.numeric(gradient) && is.vector(gradient)) {
+            attr(out, "gradient") <- gradient[p_ind]
+          }
         }
-        if (inherits(out, "hessian")) {
-          attr(out, "hessian") <- attr(out, "hessian")[p_ind, p_ind]
+        if ("hessian" %in% names(attributes(out))) {
+          hessian <- attr(out, "hessian")
+          if (is.numeric(hessian) && is.matrix(hessian)) {
+            attr(out, "hessian") <- hessian[p_ind, p_ind, drop = FALSE]
+          }
         }
         out
       }
       f_small_out <- optimizeR::apply_optimizer(
-        optimizer = optimizer, f = f_small, p = est[p_ind], ...
+        optimizer = base_optimizer, objective = f_small, initial = est[p_ind], ...
       )
-      est[p_ind] <- f_small_out[["z"]]
-      seq <- rbind(seq, c(it, part, f_small_out$time, est))
+      est[p_ind] <- f_small_out[["parameter"]]
+      seq <- rbind(seq, c(it, part, f_small_out[["seconds"]], est))
       if (nrow(seq) > length(partition)) {
         curr <- as.numeric(seq[nrow(seq) - length(partition), -(1:3)])
         last <- as.numeric(seq[nrow(seq), -(1:3)])